requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
requests==2.31.0
Flask-WTF>=1.2.0
Flask-Login>=0.6.0
Flask-Talisman>=1.0.0
Flask-Limiter>=3.0.0
bleach>=6.0.0
cryptography>=41.0.0
